POA: Áreas (reguistro, recurso, evaluacion, justificacion) Finalizado Validaciones (justificacion, recurso, final) Finalizado Evaluacion PyR Finalizado Perosnal: Incidencias e Incapacidades Finalizados Fotos personal Finalizado

+++ src/main/java/mx/edu/utxj/pye/sgi/entity/ch/EventosAreas.java
@@ -42,7 +42,7 @@ public class EventosAreas implements Serializable {
         this.eventosAreasPK = eventosAreasPK;
     }
 
-    public EventosAreas(int evento, int areaOperativa) {
+    public EventosAreas(int evento, short areaOperativa) {
         this.eventosAreasPK = new EventosAreasPK(evento, areaOperativa);
     }
